chore: refactor singletons
