+++ src/core/Engine.ts
@@ -3,7 +3,6 @@ import { EventManager, EventCallback } from './EventManager'
 
 /**
  * The Engine keeps track of all ECS-related entities and systems.
- * It a singleton and should only be lazily instanced via the GetInstance() method.
  * @fires entity_added
  * @fires entity_removed
  * @fires system_added
@@ -16,16 +15,7 @@ export class Engine {
   private eventManager: EventManager = new EventManager()
   private pool: Entity[] = []
 
-  private constructor() {}
-
-  static GetInstance(): Engine {
-    if (!Engine.Instance) {
-      Engine.Instance = new Engine()
-    }
-    return Engine.Instance
-  }
-
-  private constructor() {}
+  constructor() {}
 
   /**
    * Adds a new System to the engine's update loop.
