+++ src/core/index.ts
@@ -1,11 +1,11 @@
 import { Engine } from './Engine'
+import { InputController } from './InputController'
 
+const engine = new Engine()
+const inputController = new InputController()
+
+export { Engine, engine, inputController }
 export { Component } from './Component'
 export { Entity } from './Entity'
 export { System } from './System'
-export { InputController, Key } from './InputController'
-
-export {
-  engine: new Engine(),
-  Engine
-}
+export { Key } from './InputController'
